+++ examples/upload/src/application.ex
@@ -1,0 +1,9 @@
+defmodule Upload.Application do
+  use Application
+
+  def start(_type, _args) do
+    children = []
+    opts = [strategy: :one_for_one, name: Upload.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+end